+++ src/covid_shared/paths.py
@@ -4,11 +4,9 @@ from datetime import datetime
 
 # Input data paths
 JOHNS_HOPKINS_REPO = 'https://github.com/CSSEGISandData/COVID-19/archive/master.zip'
-ITALY_REPO = 'https://github.com/pcm-dpc/COVID-19/archive/master.zip'
 DESCARTES_REPO = 'https://github.com/descarteslabs/DL-COVID-19/archive/master.zip'
 NY_TIMES_REPO = 'https://github.com/nytimes/covid-19-data/archive/master.zip'
 ONEDRIVE_PATH = "covid-onedrive:'COVID-19 Resource Hub'"
-CITYMAPPER_MOBILITY_TEMPLATE = "https://cdn.citymapper.com/data/cmi/Citymapper_Mobility_Index_{DATE}.csv"
 OPEN_COVID19_GROUP_REPO = "https://github.com/beoutbreakprepared/nCoV2019/archive/master.zip"
 CDC_DEATHS_BY_RACE_ETHNICITY_AGE_STATE = "https://data.cdc.gov/api/views/ks3g-spdg/rows.csv?accessType=DOWNLOAD"
 
@@ -17,8 +15,6 @@ DATA_INTAKE_J_DIR = Path('/home/j/Project/covid/data_intake')
 SEROLOGY_DATA = DATA_INTAKE_J_DIR / 'serology'
 SEROSURVEY_SUPPLEMENTAL_DATA = SEROLOGY_DATA / 'supplemental_serosurvey_metadata'
 WANING_IMMUNITY_DATA = SEROLOGY_DATA / 'waning_immunity'
-US_SYMPTOM_SURVEY_DATA = DATA_INTAKE_J_DIR / 'symptom_survey/us'
-GLOBAL_SYMPTOM_SURVEY_DATA = DATA_INTAKE_J_DIR / 'symptom_survey/global'
 PULSE_SURVEY_DATA = DATA_INTAKE_J_DIR / 'pulse_survey/US Census_pulse surveys'
 SOCIAL_DISTANCING_DATA = DATA_INTAKE_J_DIR / 'social distancing'
 CFR_AGE_MAX_PLANCK_DATA = DATA_INTAKE_J_DIR / 'CFR-age/MaxPlanck'
@@ -27,55 +23,80 @@ VACCINE_DATA = DATA_INTAKE_J_DIR / 'vaccine'
 MEXICO_CFR_AGE = DATA_INTAKE_J_DIR / 'data' / 'out' / 'cfr-age' / 'mexico' / 'latest'
 MEXICO_HFR_AGE = DATA_INTAKE_J_DIR / 'data' / 'out' / 'hospitalization-age' / 'mexico' / 'latest'
 VARIANTS_DATA_GISAID = DATA_INTAKE_J_DIR / 'GISAID'
-VARIANTS_DATA_CDC = DATA_INTAKE_J_DIR / 'variants' / 'CDC'
-VARIANTS_DATA_DENMARK = DATA_INTAKE_J_DIR / 'variants' / 'Denmark'
+VARIANTS_DATA = DATA_INTAKE_J_DIR / 'variants'
+GOOGLE_MOBILITY_DATA = DATA_INTAKE_J_DIR / 'mobility' / 'google'
 
-# Shared paths
+##################
+# Executor paths #
+##################
 EXEC_R_SCRIPT_PATH = Path('/share/singularity-images/lbd/shells/singR.sh')
 R_SINGULARITY_IMAGE_PATH = Path('/ihme/singularity-images/lbd/releases/lbd_full_20200128.simg')
 
-RCLONE_CONFIG_PATH = Path('/ihme/covid-19/.config/rclone/rclone.conf')
+################
+# Shared paths #
+################
+# Concrete share drive root
+COVID_19_2 = Path('/ihme/covid-19-2')
+ARCHIVE_ROOT = COVID_19_2 / 'archive'
 
-UNVERSIONED_INPUTS_ROOT = Path('/ihme/covid-19/unversioned-inputs')
-STATIC_DATA_INPUTS_ROOT = Path('/ihme/covid-19/static-data')
-SNAPSHOT_ROOT = Path('/ihme/covid-19/snapshot-data/')
-MODEL_INPUTS_ROOT = Path('/ihme/covid-19/model-inputs/')
-VISIT_VOLUME_ROOT = Path('/ihme/covid-19/visit-volume/')
-MORTALITY_RATIO_ROOT = Path('/ihme/covid-19/mortality-ratio')
-INFECTION_FATALITY_RATIO_ROOT = Path('/ihme/covid-19/infection-fatality-ratio')
-INFECTION_HOSPITALIZATION_RATIO_ROOT = Path('/ihme/covid-19/infection-hospitalization-ratio')
-INFECTION_DETECTION_RATE_ROOT = Path('/ihme/covid-19/infection-detection-rate')
-HOSPITAL_DEATH_RATIO_ROOT = Path('/ihme/covid-19/hospitalization-death-ratio')
-DEATHS_SPLINE_OUTPUT_ROOT = Path('/ihme/covid-19/deaths-outputs/')
-PAST_INFECTIONS_ROOT = Path('/ihme/covid-19/past-infections')
-INFECTIONATOR_OUTPUTS = Path('/ihme/covid-19/seir-inputs')
-SHAPEFILE_ROOT = Path('/ihme/covid-19/shapefiles')
-WEBSCRAPER_ROOT = Path('/ihme/covid-19/webscrape')
+# Main share drive root. This is a symlink farm for COVID_19_2.
+# All paths should point here.
+COVID_19 = Path('/ihme/covid-19')
 
-VARIANT_OUTPUT_ROOT = Path('/ihme/covid-19/variant-scaleup')
-MASK_USE_OUTPUT_ROOT = Path('/ihme/covid-19/mask-use-outputs')
-PNEUMONIA_OUTPUT_ROOT = Path('/ihme/covid-19/pneumonia')
-POPULATION_DENSITY_OUTPUT_ROOT = Path('/ihme/covid-19/population-density')
-MOBILITY_COVARIATES_OUTPUT_ROOT = Path('/ihme/covid-19/mobility-covariate')
+# Shared config for running rclone on the IHME OneDrive
+RCLONE_CONFIG_PATH = COVID_19 / '.config' / 'rclone' / 'rclone.conf'
+
+# Top level directories.  These represent, mostly, outputs of pipeline stages.
+# Keep them in alphabetical order.
+# TODO: This should definitely be in a container.
+AGE_SPECIFIC_RATES_ROOT = COVID_19 / 'age-specific-rates'
+DURATIONS_ROOT = COVID_19 / 'durations'
+EXCESS_MORTALITY_ROOT = COVID_19 / 'excess-mortality'
+GBD_ROOT = COVID_19 / 'gbd'
+HISTORICAL_MODEL_ROOT = COVID_19 / 'historical-model'
+MANDATES_ROOT = COVID_19 / 'mandates'
+MASK_USE_OUTPUT_ROOT = COVID_19 / 'mask-use-outputs'
+MOBILITY_COVARIATES_OUTPUT_ROOT = COVID_19 / 'mobility-covariate'
 MOBILITY_COVARIATES_GPR_OUTPUT_ROOT = MOBILITY_COVARIATES_OUTPUT_ROOT / 'gpr_outputs'
-TESTING_OUTPUT_ROOT = Path('/ihme/covid-19/testing-outputs')
-VACCINE_COVERAGE_OUTPUT_ROOT = Path('/ihme/covid-19/vaccine-coverage')
-SEIR_COVARIATES_OUTPUT_ROOT = Path('/ihme/covid-19/seir-covariates')
+MODEL_INPUTS_ROOT = COVID_19 / 'model-inputs'
+MORTALITY_AGE_PATTERN_ROOT = COVID_19 / 'mortality-age-pattern'
+MORTALITY_RATIO_ROOT = COVID_19 / 'mortality-ratio'
+PAST_INFECTIONS_ROOT = COVID_19 / 'past-infections'
+PNEUMONIA_OUTPUT_ROOT = COVID_19 / 'pneumonia'
+POPULATION_DENSITY_OUTPUT_ROOT = COVID_19 / 'population-density'
+RESULTRON_ROOT = COVID_19 / 'resultron'
+SEIR_COVARIATE_PRIORS_ROOT = COVID_19 / 'seir-covariate-priors'
+SEIR_COVARIATES_OUTPUT_ROOT = COVID_19 / 'seir-covariates'
+SEIR_DIAGNOSTICS_OUTPUTS = COVID_19 / 'seir-diagnostics'
+SEIR_FORECAST_OUTPUTS = COVID_19 / 'seir-forecast'
+SEIR_FINAL_OUTPUTS = COVID_19 / 'seir-outputs'
+SEIR_REGRESSION_OUTPUTS = COVID_19 / 'seir-regression'
+SHAPEFILE_ROOT = COVID_19 / 'shapefiles'
+SNAPSHOT_ROOT = COVID_19 / 'snapshot-data'
+STATIC_DATA_INPUTS_ROOT = COVID_19 / 'static-data'
+SYMPTOM_SURVEY_ROOT = COVID_19 / 'symptom-survey-data'
+TESTING_OUTPUT_ROOT = COVID_19 / 'testing-outputs'
+UNVERSIONED_INPUTS_ROOT = COVID_19 / 'unversioned-inputs'
+VACCINE_COVERAGE_OUTPUT_ROOT = COVID_19 / 'vaccine-coverage'
+VARIANT_OUTPUT_ROOT = COVID_19 / 'variant-scaleup'
+VISIT_VOLUME_ROOT = COVID_19 / 'visit-volume/'
+WEBSCRAPER_ROOT = COVID_19 / 'webscrape'
 
-SEIR_REGRESSION_OUTPUTS = Path('/ihme/covid-19/seir-regression')
-SEIR_FORECAST_OUTPUTS = Path('/ihme/covid-19/seir-forecast')
-SEIR_FINAL_OUTPUTS = Path('/ihme/covid-19/seir-outputs')
-SEIR_DIAGNOSTICS_OUTPUTS = Path('/ihme/covid-19/seir-diagnostics')
-
-
-# Shared file names
+######################################
+# Shared file and subdirectory names #
+######################################
 METADATA_FILE_NAME = Path('metadata.yaml')
+
+LOG_DIR = Path("logs")
+LOG_FILE_NAME = Path("master_log.txt")
+DETAILED_LOG_FILE_NAME = Path("master_log.json")
+
 BEST_LINK = Path('best')
 LATEST_LINK = Path('latest')
 PRODUCTION_RUN = Path('production-runs')
 
+# ???
 JOHNS_HOPKINS_OUTPUT_DIR_NAME = Path('johns_hopkins_repo')
-ITALY_OUTPUT_DIR_NAME = Path('italy_repo')
 NY_TIMES_OUTPUT_DIR_NAME = Path('ny_times_repo')
 CDC_OUTPUT_DIR_NAME = Path('cdc_data')
 MOBILITY_OUTPUT_DIR_NAME = Path('mobility_data')
@@ -84,8 +105,6 @@ OPEN_COVID19_OUTPUT_DIR_NAME = Path('open_covid19_working_group')
 SEROSURVEY_OUTPUT_DIR_NAME = Path('serosurvey_data')
 SEROSURVEY_SUPPLEMENTAL_OUTPUT_DIR_NAME = Path('supplemental_serosurvey_metadata')
 SYMPTOM_SURVEY_OUTPUT_DIR_NAME = Path('symptom_survey')
-US_SYMPTOM_SURVEY_OUTPUT_DIR_NAME = Path('us')
-GLOBAL_SYMPTOM_SURVEY_OUTPUT_DIR_NAME = Path('global')
 PULSE_SURVEY_OUTPUT_DIR_NAME = Path('pulse_survey')
 PULSE_SURVEY_US_CENSUS_OUTPUT_DIR_NAME = Path('US Census_pulse surveys')
 SOCIAL_DISTANCING_DIR_NAME = Path('social distancing')
@@ -96,14 +115,8 @@ VACCINE_DIR_NAME = Path('vaccine')
 MEXICO_CFR_AGE_DIR_NAME = Path('mexico_cfr_age')
 MEXICO_HFR_AGE_DIR_NAME = Path('mexico_hfr_age')
 VARIANTS_DIR_NAME = Path('Variants')
-VARIANTS_CDC_DIR_NAME = VARIANTS_DIR_NAME / 'CDC'
-VARIANTS_DENMARK_DIR_NAME = VARIANTS_DIR_NAME / 'Denmark'
 SEROLOGY_DIR_NAME = Path('serology')
 WANING_IMMUNITY_DIR_NAME = Path('waning_immunity')
-
-LOG_DIR = Path("logs")
-LOG_FILE_NAME = Path("master_log.txt")
-DETAILED_LOG_FILE_NAME = Path("master_log.json")
 
 
 def latest_production_snapshot_path():